Remove futures context (#1941)
* Remove futures context

* Bring back futures data query

---------
