+++ sections/futures/OrderSizing/OrderSizeSlider.tsx
@@ -6,13 +6,13 @@ import styled from 'styled-components';
 import ErrorView from 'components/ErrorView';
 import { FlexDivRow } from 'components/layout/flex';
 import StyledSlider from 'components/Slider/StyledSlider';
-import { useFuturesContext } from 'contexts/FuturesContext';
 import { editCrossMarginSize } from 'state/futures/actions';
 import {
 	selectAboveMaxLeverage,
 	selectCrossMarginBalanceInfo,
 	selectLeverageSide,
 	selectMaxLeverage,
+	selectMaxUsdInputAmount,
 	selectPosition,
 	selectTradeSizeInputs,
 } from 'state/futures/selectors';
@@ -20,7 +20,6 @@ import { useAppDispatch, useAppSelector } from 'state/hooks';
 
 export default function OrderSizeSlider() {
 	const { t } = useTranslation();
-	const { maxUsdInputAmount } = useFuturesContext();
 	const dispatch = useAppDispatch();
 	const { freeMargin: freeCrossMargin } = useAppSelector(selectCrossMarginBalanceInfo);
 	const { susdSizeString } = useAppSelector(selectTradeSizeInputs);
@@ -28,6 +27,7 @@ export default function OrderSizeSlider() {
 	const maxLeverage = useAppSelector(selectMaxLeverage);
 	const leverageSide = useAppSelector(selectLeverageSide);
 	const position = useAppSelector(selectPosition);
+	const maxUsdInputAmount = useAppSelector(selectMaxUsdInputAmount);
 
 	const [percent, setPercent] = useState(0);
 	const [usdValue, setUsdValue] = useState(susdSizeString);
